requirements.txt
absl-py==1.4.0
appdirs==1.4.4
asttokens==2.4.1
av==12.1.0
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
datrie==0.8.2
decorator==5.1.1
dill==0.3.8
dm-tree==0.1.8
docker-pycreds==0.4.0
executing==2.0.1
filelock==3.15.3
fonttools==4.53.0
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.5.1
html2image==2.0.4.3
html5lib==1.1
huggingface-hub==0.23.4
iceberg-dsl @ git+https://github.com/revalo/iceberg.git@db446c1ed6c44d2a10adf059eb21bfceaf534f0c
idna==3.7
imgkit==1.2.3
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.5
lark==1.1.9
lxml==5.2.2
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
packaging==24.1
parso==0.8.4
pathos==0.3.2
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.3.0
pox==0.3.4
ppft==1.7.6.8
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.12.0
pycparser==2.22
pydyf==0.10.0
Pygments==2.18.0
pyparsing==3.1.2
pyphen==0.15.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scipy==1.13.0
sentry-sdk==2.6.0
setproctitle==1.3.3
six==1.16.0
skia-python==87.5
smmap==5.0.1
stack-data==0.6.3
sympy==1.12.1
timm==1.0.3
tinycss2==1.3.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wandb==0.16.4
wcwidth==0.2.13
weasyprint==62.2
webencodings==0.5.1
websocket-client==1.8.0
zopfli==0.2.3
